fix: consolidate duplicated helper functions and add shared utilities
Critical issues addressed:
1. formatBytes() - consolidated from 4 locations to 1 canonical version in client-safe.ts
2. File security functions - consolidated DANGEROUS_MIME_TYPES and isDangerousMimeType into packages/lib
3. getClientIP() - now exported from @pagespace/lib/server and used in 15 API routes
4. MCP tool converter - created shared module in packages/lib for web app (desktop keeps own version due to zod v3/v4 incompatibility)

Changes:
- packages/lib: Added zod dependency, new mcp/ module with shared types and converter
- packages/lib: Export getClientIP from server.ts
- apps/web: Updated 15 API routes to use getClientIP()
- apps/web: Updated mcp-tool-converter to re-export from @pagespace/lib
- apps/processor: Import file security utilities from @pagespace/lib
