+++ apps/web/src/app/api/auth/mobile/login/route.ts
@@ -9,6 +9,7 @@ import {
   RATE_LIMIT_CONFIGS,
   decodeToken,
   validateOrCreateDeviceToken,
+  getClientIP,
 } from '@pagespace/lib/server';
 import { generateCSRFToken, getSessionIdFromJWT } from '@pagespace/lib/server';
 import { loggers, logAuthEvent } from '@pagespace/lib/server';
@@ -38,9 +39,7 @@ export async function POST(req: Request) {
     const { email, password, deviceId, platform, deviceName, appVersion, deviceToken: existingDeviceToken } = validation.data;
 
     // Rate limiting by IP address and email
-    const clientIP = req.headers.get('x-forwarded-for')?.split(',')[0] ||
-                     req.headers.get('x-real-ip') ||
-                     'unknown';
+    const clientIP = getClientIP(req);
 
     const ipRateLimit = checkRateLimit(clientIP, RATE_LIMIT_CONFIGS.LOGIN);
     const emailRateLimit = checkRateLimit(email.toLowerCase(), RATE_LIMIT_CONFIGS.LOGIN);
