+++ apps/web/src/app/dashboard/storage/page.tsx
@@ -36,6 +36,7 @@ import {
 import { formatDistanceToNow } from "date-fns";
 import { toast } from "sonner";
 import { fetchWithAuth } from "@/lib/auth/auth-fetch";
+import { formatBytes } from "@pagespace/lib/client-safe";
 
 interface StorageInfo {
   quota: {
@@ -98,13 +99,6 @@ const getFileIcon = (mimeType: string) => {
   if (mimeType.includes('presentation') || mimeType.includes('powerpoint')) return Presentation;
   if (mimeType.includes('zip') || mimeType.includes('compress')) return Archive;
   return File;
-};
-
-const formatBytes = (bytes: number): string => {
-  const sizes = ['B', 'KB', 'MB', 'GB', 'TB'];
-  if (bytes === 0) return '0 B';
-  const i = Math.floor(Math.log(bytes) / Math.log(1024));
-  return `${Math.round(bytes / Math.pow(1024, i) * 100) / 100} ${sizes[i]}`;
 };
 
 export default function StorageDashboard() {
