+++ packages/lib/src/mcp/index.ts
@@ -1,0 +1,7 @@
+/**
+ * MCP (Model Context Protocol) utilities
+ * Shared across web and desktop packages
+ */
+
+export * from './mcp-types';
+export * from './mcp-tool-converter';
