+++ packages/lib/src/services/subscription-utils.ts
@@ -90,12 +90,5 @@ export function subscriptionAllows(subscriptionTier: SubscriptionTier, feature: 
   return config.features.includes(feature);
 }
 
-/**
- * Format bytes to human-readable string
- */
-export function formatBytes(bytes: number): string {
-  const sizes = ['B', 'KB', 'MB', 'GB', 'TB'];
-  if (bytes === 0) return '0 B';
-  const i = Math.floor(Math.log(bytes) / Math.log(1024));
-  return `${Math.round(bytes / Math.pow(1024, i) * 100) / 100} ${sizes[i]}`;
-}
+// Re-export for backward compatibility (canonical version is in client-safe.ts)
+export { formatBytes } from '../client-safe';
